feat/player : code seperation
